refactor: made hook to get privilage users
